Changing format flag in printf used in tests code.
In <odreex/ample/test.hh> test framework code, the printf function is used
with a "[%06zu]" format string containing the z flag which is an ISO C99
compliant option, but not an ISO C++11 one. Therefore and in order to continue
to use std::size_t portably with printf, switching to casting to unsigned long
long and deploying the [%06llu] format string is one of the options considered
for resolving this issue and continue using printf (for the time being...).

Thanks to Jose Diego Ferreira Martins <jose.diego@gmail.com> for reporting
this while compiling with mingw (4.7.2) on Microsoft Windows.

+++ odreex/ample/test.hh
@@ -35,16 +35,16 @@
 
 #ifdef USE_ANSI_COLORS
 #define ample_printf_(i, sz, N) \
-    printf( "\033[1;37m[%06zu]\033[0m: %s | %s |: \033[34m%s%s\033[0m\n" \
-          , i \
+    printf( "\033[1;37m[%06llu]\033[0m: %s | %s |: \033[34m%s%s\033[0m\n" \
+          , static_cast<unsigned long long>(i) \
           , result[ i ] ? "\033[36mpass\033[0m" : "\033[31mfail\033[0m" \
           , output[ i ] ? "\033[1;36mpass\033[0m" : "\033[1;31mfail\033[0m" \
           , cbuf \
           , sz < N ? "" : "..." )
 #else
 #define ample_printf_(i, sz, N) \
-    printf( "[%06zu]: %s | %s |: %s%s\n" \
-          , i \
+    printf( "[%06llu]: %s | %s |: %s%s\n" \
+          , static_cast<unsigned long long>(i) \
           , result[ i ] ? "pass" : "fail" \
           , output[ i ] ? "pass" : "fail" \
           , cbuf \
